+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Settings for hiding UI elements related to RPE and TUT
 - Option to prefer exercise in training session
 - Shortcut for inserting values of previous set into current set
+- Suggestion of exercises that train similar muscles when replacing exercise in training session
 
 ### Changed